+++ mod.ts
@@ -2,12 +2,13 @@ export type {
 	ArkiveManifest,
 	BlockHandler,
 	EventHandlerFor,
-} from "./src/arkiver/types.ts";
-export { Manifest } from "./src/arkiver/manifest-builder.ts";
-export { logger } from "./src/logger.ts";
-export { Store } from "./src/arkiver/store.ts";
-export { Arkiver } from "./src/arkiver/arkiver.ts";
-export { createEntity } from "./src/graphql/entity.ts";
-export { Types } from "./src/deps.ts";
-export { buildSchemaFromEntities } from "./src/graphql/builder.ts";
-export { supportedChains } from "./src/chains.ts";
+} from './src/arkiver/types.ts'
+export { Manifest } from './src/arkiver/manifest-builder.ts'
+export { ArkiveConsoleLogHandler, logger } from './src/logger.ts'
+export { Store } from './src/arkiver/store.ts'
+export { Arkiver } from './src/arkiver/arkiver.ts'
+export { createEntity } from './src/graphql/entity.ts'
+export { Types } from './src/deps.ts'
+export { buildSchemaFromEntities } from './src/graphql/builder.ts'
+export { supportedChains } from './src/chains.ts'
+export { defaultArkiveData } from './src/utils.ts'