+++ src/logger.ts
@@ -1,14 +1,23 @@
-import { log } from './deps.ts'
-log.setup({
-	handlers: {
-		console: new log.handlers.ConsoleHandler('DEBUG'),
-	},
-	loggers: {
-		default: {
-			level: 'DEBUG',
-			handlers: ['console'],
-		},
-	},
-})
+import { colors, ConsoleHandler, log } from './deps.ts'
 
-export const logger = log.getLogger()
+export class ArkiveConsoleLogHandler extends ConsoleHandler {
+	private arkiveName: string
+
+	constructor(
+		levelName: log.LevelName,
+		options: log.HandlerOptions & { arkiveName: string },
+	) {
+		super(levelName, options)
+		this.arkiveName = options.arkiveName
+	}
+
+	override format(logRecord: log.LogRecord): string {
+		let msg = super.format(logRecord)
+
+		msg = `${colors.blue(`[${this.arkiveName}]`)} ${msg}`
+
+		return msg
+	}
+}
+
+export const logger = () => log.getLogger('arkiver')